+++ src/extensions/mvs/tree/mvs/param-types.ts
@@ -25,6 +25,9 @@ export const ParseFormatT = literal(
     'lammpstrj', // + coordinates
     // coordinates
     'xtc',
+    'nctraj',
+    'dcd',
+    'trr',
     // topology
     'psf',
     'prmtop',
@@ -50,6 +53,9 @@ export const MolstarParseFormatT = literal(
     'lammpstrj',
     // coordinates
     'xtc',
+    'nctraj',
+    'dcd',
+    'trr',
     // topology
     'psf',
     'prmtop',