+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - MolViewSpec
   - Add support for "topology" formats (TOP, PRMTOP, PSF)
   - Add support for additional "coordiates" formats (NCTRAJ, DCD, TRR)
+  - Fix coarse structure selection
 
 ## [v5.2.0] - 2025-10-31
 - Handle transparency updates on ImagePass