+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Note that since we don't clearly distinguish between a public and private interf
 ## [Unreleased]
 - Update loading message in MVS Stories Viewer
 - Add `Canvas3D.setAttribs`
+- Fix `normalizeWheel` "spin" calculation fallback
 - MolViewSpec
   - Add support for "topology" formats (TOP, PRMTOP, PSF)
   - Add support for additional "coordiates" formats (NCTRAJ, DCD, TRR)